# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Midgard Public API
  version: 2.11.4
  contact:
    email: devs@thorchain.org
  description: >-
    The Midgard Public API queries THORChain and any chains linked via the Bifröst
    and prepares information about the network to be readily available for public users.
    The API parses transaction event data from THORChain and stores them
    in a time-series database to make time-dependent queries easy.
    Midgard does not hold critical information.
    To interact with BEPSwap and Asgardex, users should query THORChain directly.
paths:
...
  "/v2/pools":
    get:
      operationId: GetPools
      summary: Pools List
      description: Returns an array containing details for a set of pools
      parameters:
        - name: status
          in: query
          description: Filter for only pools with this status
          required: false
          schema:
            type: string
            enum: ["available", "staged", "suspended"]
        - name: period
          in: query
          description: |
              Specifies the base interval from which annualPercentageRate and poolAPY is extrapolated.
              Default is 30d.
          required: false
          example: "100d"
          schema:
              type: string
              enum: ["1h", "24h", "7d", "30d", "90d", "100d", "180d", "365d", "all"]
      responses:
        "200":
          $ref: '#/components/responses/PoolsResponse'
  "/v2/pool/{asset}":
    get:
      operationId: GetPool
      summary: Details of a Pool
      description: 'Returns details of the pool: depths, price, 24h volume, APY. '
      parameters:
        - in: path
          name: asset
          description: pool name
          required: true
          schema:
            type: string
          example: 'BNB.TOMOB-1E1'
        - name: period
          in: query
          description: |
              Specifies the base interval from which APY is extrapolated.
              Default is 30d.
          required: false
          example: "100d"
          schema:
              type: string
              enum: ["1h", "24h", "7d", "30d", "90d", "100d", "180d", "365d", "all"]
      responses:
        "200":
          $ref: '#/components/responses/PoolResponse'
  "/v2/pool/{asset}/stats":
    get:
      operationId: GetPoolStats
      summary: Pool Statistics
      description: |
        Statistics about the pool. The description of the fields have pointers about the
        corresponding v2/history location. Visit the history endpoint for drilldowns.
      parameters:
        - in: path
          name: asset
          description: pool name
          required: true
          schema:
            type: string
          example: 'BNB.TOMOB-1E1'
        - name: period
          in: query
          description: |
            Restricts aggregation type fields to the last period only.
            Default is 30d.
          required: false
          example: "24h"
          schema:
            type: string
            enum: ["1h", "24h", "7d", "30d", "90d", "100d", "180d", "365d", "all"]
      responses:
        "200":
          $ref: '#/components/responses/PoolStatsResponse'
  "/v2/history/depths/{pool}":
    get:
      operationId: GetDepthHistory
      summary: Depth and Price History
      description: |
        Returns the asset and rune depths and price.
        The values report the state at the end of each interval.

        History endpoint has two modes:
        * With Interval parameter it returns a series of time buckets. From and To dates will
          be rounded to the Interval boundaries.
        * Without Interval parameter a single From..To search is performed with exact timestamps.


        * Interval: possible values: 5min, hour, day, week, month, quarter, year.
        * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
        * from/to: optional int, unix second.

        Possible usages with interval.
        * last 10 days: `?interval=day&count=10`
        * last 10 days before to: `?interval=day&count=10&to=1608825600`
        * next 10 days after from: `?interval=day&count=10&from=1606780800`
        * Days between from and to. From defaults to start of chain, to defaults to now.
          Only the first 400 intervals are returned:
          `interval=day&from=1606780800&to=1608825600`

        Pagination is possible with from&count and then using the returned meta.endTime as the
        From parameter of the next query.

        Possible configurations without interval:
        * exact search for one time frame: `?from=1606780899&to=1608825600`
        * one time frame until now: `?from=1606780899`
        * from chain start until now: no query parameters
      parameters:
        - name: pool
          in: path
          description: Return stats for this single pool.
          required: true
          schema:
            type: string
        - name: interval
          in: query
          description: Interval of calculations
          required: false
          example: "day"
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "quarter", "year"]
        - name: count
          in: query
          description: Number of intervals to return. Should be between [1..400].
          required: false
          example: 30
          schema:
            type: integer
        - name: to
          in: query
          description: |
            End time of the query as unix timestamp. If only count is given, defaults to now.
          required: false
          example: 1608825600
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          description: Start time of the query as unix timestamp
          required: false
          example: 1606780800
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/DepthHistoryResponse'
  "/v2/history/ohlcv/{pool}":
    get:
      operationId: GetOHLCVHistory
      summary: Get historical OHLCV (Open, High, Low, Close, Volume) of a pool
      description: |
        Returns the asset and rune depths and price.
        The values report the state at the end of each interval.

        History endpoint has two modes:
        * With Interval parameter it returns a series of time buckets. From and To dates will
          be rounded to the Interval boundaries.
        * Without Interval parameter a single From..To search is performed with exact timestamps.


        * Interval: possible values: 5min, hour, day, week, month, quarter, year.
        * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
        * from/to: optional int, unix second.

        Possible usages with interval.
        * last 10 days: `?interval=day&count=10`
        * last 10 days before to: `?interval=day&count=10&to=1608825600`
        * next 10 days after from: `?interval=day&count=10&from=1606780800`
        * Days between from and to. From defaults to start of chain, to defaults to now.
          Only the first 400 intervals are returned:
          `interval=day&from=1606780800&to=1608825600`

        Pagination is possible with from&count and then using the returned meta.endTime as the
        From parameter of the next query.

        Possible configurations without interval:
        * exact search for one time frame: `?from=1606780899&to=1608825600`
        * one time frame until now: `?from=1606780899`
        * from chain start until now: no query parameters
      parameters:
        - name: pool
          in: path
          description: Return stats for this single pool.
          required: true
          schema:
            type: string
        - name: interval
          in: query
          description: Interval of calculations
          required: false
          example: "day"
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "quarter", "year"]
        - name: count
          in: query
          description: Number of intervals to return. Should be between [1..400].
          required: false
          example: 30
          schema:
            type: integer
        - name: to
          in: query
          description: |
            End time of the query as unix timestamp. If only count is given, defaults to now.
          required: false
          example: 1608825600
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          description: Start time of the query as unix timestamp
          required: false
          example: 1606780800
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/OHLCVHistoryResponse'
  "/v2/history/earnings":
    get:
      operationId: GetEarningsHistory
      summary: Earnings History
      description: |
        Returns earnings data for the specified interval.

        History endpoint has two modes:
        * With Interval parameter it returns a series of time buckets. From and To dates will
          be rounded to the Interval boundaries.
        * Without Interval parameter a single From..To search is performed with exact timestamps.


        * Interval: possible values: 5min, hour, day, week, month, quarter, year.
        * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
        * from/to: optional int, unix second.

        Possible usages with interval.
        * last 10 days: `?interval=day&count=10`
        * last 10 days before to: `?interval=day&count=10&to=1608825600`
        * next 10 days after from: `?interval=day&count=10&from=1606780800`
        * Days between from and to. From defaults to start of chain, to defaults to now.
          Only the first 400 intervals are returned:
          `interval=day&from=1606780800&to=1608825600`

        Pagination is possible with from&count and then using the returned meta.endTime as the
        From parameter of the next query.

        Possible configurations without interval:
        * exact search for one time frame: `?from=1606780899&to=1608825600`
        * one time frame until now: `?from=1606780899`
        * from chain start until now: no query parameters
      parameters:
        - name: interval
          in: query
          description: Interval of calculations
          required: false
          example: "day"
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "quarter", "year"]
        - name: count
          in: query
          description: Number of intervals to return. Should be between [1..400].
          required: false
          example: 30
          schema:
            type: integer
        - name: to
          in: query
          description: |
            End time of the query as unix timestamp. If only count is given, defaults to now.
          required: false
          example: 1608825600
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          description: Start time of the query as unix timestamp
          required: false
          example: 1606780800
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/EarningsHistoryResponse'
  "/v2/history/ts-swaps":
    get:
      operationId: GetTSSwapHistory
      summary: THORSwap Swaps History
      description: |
        Returns swap count, volume, fees, slip in specified interval routed by THORSwap.
        If pool is not specified returns for all pools

        History endpoint has two modes:
        * With Interval parameter it returns a series of time buckets. From and To dates will
          be rounded to the Interval boundaries.
        * Without Interval parameter a single From..To search is performed with exact timestamps.


        * Interval: possible values: 5min, hour, day, week, month, quarter, year.
        * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
        * from/to: optional int, unix second.

        Possible usages with interval.
        * last 10 days: `?interval=day&count=10`
        * last 10 days before to: `?interval=day&count=10&to=1608825600`
        * next 10 days after from: `?interval=day&count=10&from=1606780800`
        * Days between from and to. From defaults to start of chain, to defaults to now.
          Only the first 400 intervals are returned:
          `interval=day&from=1606780800&to=1608825600`

        Pagination is possible with from&count and then using the returned meta.endTime as the
        From parameter of the next query.

        Possible configurations without interval:
        * exact search for one time frame: `?from=1606780899&to=1608825600`
        * one time frame until now: `?from=1606780899`
        * from chain start until now: no query parameters
      parameters:
        - name: pool
          in: query
          description: Return history given pool. Returns sum of all pools if missing.
          required: false
          schema:
            type: string
        - name: interval
          in: query
          description: Interval of calculations
          required: false
          example: "day"
          schema:
            type: string
            enum: [ "5min", "hour", "day", "week", "month", "quarter", "year" ]
        - name: count
          in: query
          description: Number of intervals to return. Should be between [1..400].
          required: false
          example: 30
          schema:
            type: integer
        - name: to
          in: query
          description: |
            End time of the query as unix timestamp. If only count is given, defaults to now.
          required: false
          example: 1608825600
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          description: Start time of the query as unix timestamp
          required: false
          example: 1606780800
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/SwapHistoryResponse'
  "/v2/history/swaps":
    get:
      operationId: GetSwapHistory
      summary: Swaps History
      description: |
        Returns swap count, volume, fees, slip in specified interval.
        If pool is not specified returns for all pools

        History endpoint has two modes:
        * With Interval parameter it returns a series of time buckets. From and To dates will
          be rounded to the Interval boundaries.
        * Without Interval parameter a single From..To search is performed with exact timestamps.


        * Interval: possible values: 5min, hour, day, week, month, quarter, year.
        * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
        * from/to: optional int, unix second.

        Possible usages with interval.
        * last 10 days: `?interval=day&count=10`
        * last 10 days before to: `?interval=day&count=10&to=1608825600`
        * next 10 days after from: `?interval=day&count=10&from=1606780800`
        * Days between from and to. From defaults to start of chain, to defaults to now.
          Only the first 400 intervals are returned:
          `interval=day&from=1606780800&to=1608825600`

        Pagination is possible with from&count and then using the returned meta.endTime as the
        From parameter of the next query.

        Possible configurations without interval:
        * exact search for one time frame: `?from=1606780899&to=1608825600`
        * one time frame until now: `?from=1606780899`
        * from chain start until now: no query parameters
      parameters:
        - name: pool
          in: query
          description: Return history given pool. Returns sum of all pools if missing.
          required: false
          schema:
            type: string
        - name: interval
          in: query
          description: Interval of calculations
          required: false
          example: "day"
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "quarter", "year"]
        - name: count
          in: query
          description: Number of intervals to return. Should be between [1..400].
          required: false
          example: 30
          schema:
            type: integer
        - name: to
          in: query
          description: |
            End time of the query as unix timestamp. If only count is given, defaults to now.
          required: false
          example: 1608825600
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          description: Start time of the query as unix timestamp
          required: false
          example: 1606780800
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/SwapHistoryResponse'
  "/v2/history/tvl":
    get:
      operationId: GetTVLHistory
      summary: Total Value Locked History
      description: |
        Returns total pool depths, total bonds, and total value locked in specified interval.

        Total Value Locked = Total Bonds + 2 * Total Pool Depths

        History endpoint has two modes:
        * With Interval parameter it returns a series of time buckets. From and To dates will
          be rounded to the Interval boundaries.
        * Without Interval parameter a single From..To search is performed with exact timestamps.

        * Interval: possible values: 5min, hour, day, week, month, quarter, year.
        * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
        * from/to: optional int, unix second.

        Possible usages with interval.
        * last 10 days: `?interval=day&count=10`
        * last 10 days before to: `?interval=day&count=10&to=1608825600`
        * next 10 days after from: `?interval=day&count=10&from=1606780800`
        * Days between from and to. From defaults to start of chain, to defaults to now.
          Only the first 400 intervals are returned:
          `interval=day&from=1606780800&to=1608825600`

        Pagination is possible with from&count and then using the returned meta.endTime as the
        From parameter of the next query.

        Possible configurations without interval:
        * exact search for one time frame: `?from=1606780899&to=1608825600`
        * one time frame until now: `?from=1606780899`
        * from chain start until now: no query parameters
      parameters:
        - name: interval
          in: query
          description: Interval of calculations
          required: false
          example: "day"
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "quarter", "year"]
        - name: count
          in: query
          description: Number of intervals to return. Should be between [1..400].
          required: false
          example: 30
          schema:
            type: integer
        - name: to
          in: query
          description: |
            End time of the query as unix timestamp. If only count is given, defaults to now.
          required: false
          example: 1608825600
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          description: Start time of the query as unix timestamp
          required: false
          example: 1606780800
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/TVLHistoryResponse'
  "/v2/history/liquidity_changes":
    get:
      operationId: GetLiquidityHistory
      summary: Liquidity Changes History
      description: |
        Returns withdrawals and deposits for given time interval.
        If pool is not specified returns for all pools

        History endpoint has two modes:
        * With Interval parameter it returns a series of time buckets. From and To dates will
          be rounded to the Interval boundaries.
        * Without Interval parameter a single From..To search is performed with exact timestamps.


        * Interval: possible values: 5min, hour, day, week, month, quarter, year.
        * count: [1..400]. Defines number of intervals. Don't provide if Interval is missing.
        * from/to: optional int, unix second.

        Possible usages with interval.
        * last 10 days: `?interval=day&count=10`
        * last 10 days before to: `?interval=day&count=10&to=1608825600`
        * next 10 days after from: `?interval=day&count=10&from=1606780800`
        * Days between from and to. From defaults to start of chain, to defaults to now.
          Only the first 400 intervals are returned:
          `interval=day&from=1606780800&to=1608825600`

        Pagination is possible with from&count and then using the returned meta.endTime as the
        From parameter of the next query.

        Possible configurations without interval:
        * exact search for one time frame: `?from=1606780899&to=1608825600`
        * one time frame until now: `?from=1606780899`
        * from chain start until now: no query parameters
      parameters:
        - name: pool
          in: query
          description: Return stats for given pool. Returns sum of all pools if missing
          required: false
          schema:
            type: string
        - name: interval
          in: query
          description: Interval of calculations
          required: false
          example: "day"
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "quarter", "year"]
        - name: count
          in: query
          description: Number of intervals to return. Should be between [1..400]
          required: false
          example: 30
          schema:
            type: integer
        - name: to
          in: query
          description: End time of the query as unix timestamp. If only count is given, defaults to now
          required: false
          example: 1608825600
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          description: Start time of the query as unix timestamp
          required: false
          example: 1606780800
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/LiquidityHistoryResponse'
...
  "/v2/members":
    get:
      operationId: GetMembersAdresses
      summary: Members List
      description: |
        Returns an array containing the addresses for all pool members.
        Addresses are only shown once. If there's both a RUNE address and an asset address
        for a member, only the RUNE address will be shown.
      parameters:
        - name: pool
          in: query
          description: Return only members present in the pool.
          required: false
          schema:
            type: string
      responses:
        "200":
          "$ref": "#/components/responses/MembersResponse"
  "/v2/full_member":
    get:
      operationId: GetFullMembersAdresses
      summary: Full Members List
      description: |
        Returns an array containing the addresses for all pool members.
        Addresses are only shown once. If there's both a RUNE address and an asset address
        for a member, only the RUNE address will be shown.
      parameters:
        - name: pool
          in: query
          description: Return only members present in the pool.
          required: false
          schema:
            type: string
      responses:
        "200":
          "$ref": "#/components/responses/FullMembersResponse"
  "/v2/member/{address}":
    get:
      operationId: GetMemberDetail
      summary: Member Details
      description: |
        Returns an array of statistics for all the liquidity providers associated with a given
        member address.
      parameters:
        - name: address
          in: path
          description: |
            Address to match liquidity providers. Either a rune or an asset address may be given.
          required: true
          schema:
            type: string
          example: 'bnb1jxfh2g85q3v0tdq56fnevx6xcxtcnhtsmcu64m'
      responses:
        "200":
          $ref: '#/components/responses/MemberDetailsResponse'
  "/v2/thorname/lookup/{name}":
    get:
      operationId: GetTHORNameDetail
      summary: THORName Details
      description: Returns an array of chains and their addresses associated with the given THORName
      parameters:
        - name: name
          in: path
          description: a THORName
          required: true
          schema:
            type: string
          example: 'thorchain'
      responses:
        "200":
          $ref: '#/components/responses/THORNameDetailsResponse'
...
components:
  responses:
...
    PoolsResponse:
      description: Array of pool details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PoolDetails'
    PoolStatsResponse:
      description: Stats for one pool.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PoolStatsDetail'
    PoolResponse:
      description: Object containing details for one pool
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PoolDetail'
    DepthHistoryResponse:
      description: Depth and price history
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DepthHistory'
    OHLCVHistoryResponse:
      description: OHLCV history
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OHLCVHistory'
    EarningsHistoryResponse:
      description: earnings history
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EarningsHistory'
    SwapHistoryResponse:
      description: Swap count, volume, fee and slip history
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SwapHistory'
...
    TVLHistoryResponse:
      description: Total pool depths, total bonds, and total value locked history
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TVLHistory'
...